Split behavior bound generated C tests
